+++ dca/tl_form.php
@@ -58,6 +58,7 @@ array_insert($GLOBALS['TL_DCA']['tl_form']['fields'], 0, array
     'anystores_categories' => array
     (
         'label'            => &$GLOBALS['TL_LANG']['tl_form']['anystores_categories'],
+        'exclude'          => true,
         'inputType'        => 'checkbox',
         'options_callback' => function()
         {