+++ languages/en/tl_form_field.php
@@ -1,0 +1,16 @@
+<?php // with ♥ and Contao
+
+/**
+ * anyStores for Contao Open Source CMS
+ *
+ * @copyright   (c) 2015 Tastaturberuf <mail@tastaturberuf.de>
+ * @author      Daniel Jahnsmüller <mail@jahnsmueller.net>
+ * @license     http://opensource.org/licenses/lgpl-3.0.html
+ * @package     anyStores
+ */
+
+
+array_insert($GLOBALS['TL_LANG']['FFL'], 0, array
+(
+    'stores' => array('anyStores list', '')
+));