+++ languages/en/tl_page.php
@@ -1,0 +1,18 @@
+<?php // with ♥ and Contao
+
+/**
+ * anyStores for Contao Open Source CMS
+ *
+ * @copyright   (c) 2015 Tastaturberuf <mail@tastaturberuf.de>
+ * @author      Daniel Jahnsmüller <mail@jahnsmueller.net>
+ * @license     http://opensource.org/licenses/lgpl-3.0.html
+ * @package     anyStores
+ */
+
+
+array_insert($GLOBALS['TL_LANG']['tl_page'], 0, array
+(
+    'anystores_sitemap' => array('Add anyStores', 'Check to integrate locations to the sitemap.'),
+    'anystores_detailPage' => array('Detail page', 'Choose a site with an anyStores detail module.'),
+    'anystores_categories' => array('Categories', 'Choose categories for the sitemap.')
+));