+++ classes/AnyStoresHooks.php
@@ -74,7 +74,7 @@ class AnyStoresHooks extends \Controller
             return;
         }
 
-        if ( Validator::isEmail($objStore->email) )
+        if ( \Validator::isEmail($objStore->email) )
         {
             $objForm->recipient .= ','.$objStore->email;
         }